requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
Pillow==10.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
click==8.1.7
Jinja2==3.1.2
